# backend/openapi-live.json
{"openapi":"3.0.0","paths":{"/api/v1/auth/register":{"post":{"operationId":"AuthController_register","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterDto"}}}},"responses":{"201":{"description":""}},"tags":["Auth"]}},"/api/v1/auth/login":{"post":{"operationId":"AuthController_login","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginDto"}}}},"responses":{"200":{"description":""}},"tags":["Auth"]}},"/api/v1/auth/refresh":{"post":{"operationId":"AuthController_refreshToken","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RefreshTokenDto"}}}},"responses":{"200":{"description":""}},"tags":["Auth"]}},"/api/v1/auth/logout":{"post":{"operationId":"AuthController_logout","parameters":[{"name":"authorization","required":true,"in":"header","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Auth"]}},"/api/v1/auth/forgot-password":{"post":{"operationId":"AuthController_forgotPassword","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ForgotPasswordDto"}}}},"responses":{"200":{"description":""}},"tags":["Auth"]}},"/api/v1/auth/reset-password":{"post":{"operationId":"AuthController_resetPassword","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResetPasswordDto"}}}},"responses":{"200":{"description":""}},"tags":["Auth"]}},"/api/v1/users/me":{"get":{"operationId":"UserController_getProfile","parameters":[],"responses":{"200":{"description":""}},"tags":["User"]},"patch":{"operationId":"UserController_updateProfile","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateProfileDto"}}}},"responses":{"200":{"description":""}},"tags":["User"]}},"/api/v1/users/me/change-password":{"patch":{"operationId":"UserController_changePassword","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChangePasswordDto"}}}},"responses":{"200":{"description":""}},"tags":["User"]}},"/api/v1/users/me/avatar":{"post":{"operationId":"UserController_uploadAvatar","parameters":[],"responses":{"201":{"description":""}},"tags":["User"]}},"/api/v1/workspaces":{"post":{"operationId":"WorkspaceController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateWorkspaceDto"}}}},"responses":{"201":{"description":""}},"tags":["Workspace"]},"get":{"operationId":"WorkspaceController_findAll","parameters":[],"responses":{"200":{"description":""}},"tags":["Workspace"]}},"/api/v1/workspaces/{id}":{"delete":{"operationId":"WorkspaceController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Workspace"]},"get":{"operationId":"WorkspaceController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Workspace"]},"patch":{"operationId":"WorkspaceController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateWorkspaceDto"}}}},"responses":{"200":{"description":""}},"tags":["Workspace"]}},"/api/v1/workspaces/{id}/invite":{"post":{"operationId":"WorkspaceController_createInvite","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/InviteMemberDto"}}}},"responses":{"201":{"description":""}},"tags":["Workspace"]}},"/api/v1/workspaces/accept-invite/{token}":{"post":{"operationId":"WorkspaceController_acceptInvite","parameters":[{"name":"token","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"201":{"description":""}},"tags":["Workspace"]}},"/api/v1/workspaces/{id}/members":{"get":{"operationId":"WorkspaceController_getMembers","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Workspace"]}},"/api/v1/workspaces/{id}/members/{userId}":{"patch":{"operationId":"WorkspaceController_changeMemberRole","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}},{"name":"userId","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChangeRoleDto"}}}},"responses":{"200":{"description":""}},"tags":["Workspace"]},"delete":{"operationId":"WorkspaceController_removeMember","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}},{"name":"userId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Workspace"]}},"/api/v1/workspaces/{id}/leave":{"delete":{"operationId":"WorkspaceController_leaveWorkspace","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Workspace"]}},"/api/v1/workspaces/{wsId}/projects":{"post":{"operationId":"ProjectController_create","parameters":[{"name":"wsId","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateProjectDto"}}}},"responses":{"201":{"description":""}},"tags":["Project"]},"get":{"operationId":"ProjectController_findAll","parameters":[{"name":"wsId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Project"]}},"/api/v1/projects/{id}":{"get":{"operationId":"ProjectController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Project"]},"patch":{"operationId":"ProjectController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateProjectDto"}}}},"responses":{"200":{"description":""}},"tags":["Project"]},"delete":{"operationId":"ProjectController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Project"]}},"/api/v1/projects/{id}/archive":{"post":{"operationId":"ProjectController_archive","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Project"]}},"/api/v1/projects/{id}/unarchive":{"post":{"operationId":"ProjectController_unarchive","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Project"]}},"/api/v1/projects/{id}/pin":{"post":{"operationId":"ProjectController_pin","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Project"]}},"/api/v1/projects/{id}/unpin":{"post":{"operationId":"ProjectController_unpin","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Project"]}},"/api/v1/projects/{projectId}/tasks":{"post":{"operationId":"TaskController_create","parameters":[{"name":"projectId","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateTaskDto"}}}},"responses":{"201":{"description":""}},"tags":["Task"]},"get":{"operationId":"TaskController_findAll","parameters":[{"name":"projectId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Task"]}},"/api/v1/tasks/{id}":{"get":{"operationId":"TaskController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Task"]},"patch":{"operationId":"TaskController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateTaskDto"}}}},"responses":{"200":{"description":""}},"tags":["Task"]},"delete":{"operationId":"TaskController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Task"]}},"/api/v1/tasks/{id}/status":{"patch":{"operationId":"TaskController_updateStatus","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateStatusDto"}}}},"responses":{"200":{"description":""}},"tags":["Task"]}},"/api/v1/tasks/{id}/assign":{"post":{"operationId":"TaskController_assignMember","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AssignTaskDto"}}}},"responses":{"201":{"description":""}},"tags":["Task"]}},"/api/v1/tasks/{id}/assign/{userId}":{"delete":{"operationId":"TaskController_unassignMember","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}},{"name":"userId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Task"]}},"/api/v1/tasks/{id}/labels":{"post":{"operationId":"TaskController_addLabel","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TaskLabelDto"}}}},"responses":{"201":{"description":""}},"tags":["Task"]}},"/api/v1/tasks/{id}/labels/{labelId}":{"delete":{"operationId":"TaskController_removeLabel","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}},{"name":"labelId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Task"]}},"/api/v1/tasks/{id}/subtasks":{"post":{"operationId":"TaskController_createSubtask","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateSubtaskDto"}}}},"responses":{"201":{"description":""}},"tags":["Task"]},"get":{"operationId":"TaskController_findSubtasks","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Task"]}},"/api/v1/subtasks/{id}/complete":{"patch":{"operationId":"TaskController_toggleSubtask","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Task"]}},"/api/v1/subtasks/{id}":{"delete":{"operationId":"TaskController_removeSubtask","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Task"]}},"/api/v1/tasks/{taskId}/comments":{"post":{"operationId":"CommentController_create","parameters":[{"name":"taskId","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateCommentDto"}}}},"responses":{"201":{"description":""}},"security":[{"bearer":[]}],"summary":"Tạo comment cho task","tags":["Comment"]}},"/api/v1/tasks/{taskId}/comments/{parentId}/reply":{"post":{"operationId":"CommentController_reply","parameters":[{"name":"taskId","required":true,"in":"path","schema":{"type":"string"}},{"name":"parentId","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateCommentDto"}}}},"responses":{"201":{"description":""}},"security":[{"bearer":[]}],"summary":"Reply comment trong task","tags":["Comment"]}},"/api/v1/notifications":{"post":{"operationId":"NotificationController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateNotificationDto"}}}},"responses":{"201":{"description":""}},"security":[{"bearer":[]}],"summary":"Tạo notification (Test)","tags":["Notification"]}}},"info":{"title":"TodoList Collaboration API","description":"API documentation cho dự án TodoList Collaboration","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"RegisterDto":{"type":"object","properties":{}},"LoginDto":{"type":"object","properties":{}},"RefreshTokenDto":{"type":"object","properties":{}},"ForgotPasswordDto":{"type":"object","properties":{}},"ResetPasswordDto":{"type":"object","properties":{}},"UpdateProfileDto":{"type":"object","properties":{}},"ChangePasswordDto":{"type":"object","properties":{}},"CreateWorkspaceDto":{"type":"object","properties":{}},"UpdateWorkspaceDto":{"type":"object","properties":{}},"InviteMemberDto":{"type":"object","properties":{}},"ChangeRoleDto":{"type":"object","properties":{}},"CreateProjectDto":{"type":"object","properties":{}},"UpdateProjectDto":{"type":"object","properties":{}},"CreateTaskDto":{"type":"object","properties":{}},"UpdateTaskDto":{"type":"object","properties":{}},"UpdateStatusDto":{"type":"object","properties":{}},"AssignTaskDto":{"type":"object","properties":{}},"TaskLabelDto":{"type":"object","properties":{}},"CreateSubtaskDto":{"type":"object","properties":{}},"CreateCommentDto":{"type":"object","properties":{"content":{"type":"string","description":"Comment content"}},"required":["content"]},"CreateNotificationDto":{"type":"object","properties":{"type":{"type":"string","enum":["TASK_ASSIGNED","TASK_UNASSIGNED","TASK_STATUS_CHANGED","TASK_DUE_SOON","TASK_OVERDUE","COMMENT_ADDED","COMMENT_REPLY","MENTIONED","WORKSPACE_INVITE","PROJECT_CREATED"],"description":"Notification type"},"title":{"type":"string","description":"Notification title"},"message":{"type":"string","description":"Notification message"},"userId":{"type":"string","description":"Target user id"}},"required":["type","title","userId"]}}}}